requirements.txt
requests==2.31.0
pyinstaller==6.3.0
pillow==10.2.0 